Update README.md
All in all good code. I see you did the accelerometer which is different then the application I chose so it is good to see different types of coding. A few small things is that I believe you are missing a ';' on line 115 right at the end, line 24 at the end and on 148. Also a common error which I found was that appcelerator could identify the "token" on a few lines, which I believe was because the variable was set up incorrectly but to be completely honest I don't know how to correct it. I still had troubles running it because of a "type 1 error" which I think is on my software rather than the code. That all I can think of for now, good job on the assignment.